+++ lib/src/positioner/positioner.dart
@@ -1,0 +1,17 @@
+import 'dart:ui';
+
+class Positioner {
+  late Offset position;
+  late Size size;
+
+  double get x => (position.dx + size.width) / 2;
+  double get y => (position.dy + size.height) / 2;
+  double get left => position.dx;
+  double get top => position.dy;
+  double get right => position.dx + size.width;
+  double get bottom => position.dy + size.height;
+  double get width => size.width;
+  double get height => size.height;
+
+  Positioner({required this.position, required this.size});
+}